Finished cell selection display
Cells now show their value on the spreadsheet panel when not highlighted, and show their contents when highligted. (this allows the user to easily edit the contents of the cell). Cell values are still displayed up top on the cell value output box. Let me know if you find any bugs.